ml-service/requirements.txt
flask
flask-cors
pandas==2.1.4
pymongo==4.6.0
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
numpy==1.26.2